Add food list screen, pagination and some ui.

+++ data/src/main/java/com/cafeapp/data/food_list/remote/repository/RemoteFoodListRepository.kt
@@ -3,5 +3,5 @@ package com.cafeapp.data.food_list.remote.repository
 import com.cafeapp.data.food_list.remote.models.FoodRemote
 
 interface RemoteFoodListRepository {
-    suspend fun getPagedFoodList(page: Int, limit: Long): List<FoodRemote>
+    suspend fun getPagedFoodList(page: Int, limit: Int): List<FoodRemote>
 }
